+++ CHANGELOG.rst
@@ -5,8 +5,36 @@ Changelog
 This project adheres to `Semantic Versioning <https://semver.org/>`_.
 
 
-2.5.0-SNAPSHOT (2021-03-18)
+2.7.0-SNAPSHOT (2021-05-11)
 ---------------------------
+
+**Added**
+
+**Fixed**
+
+**Dependencies**
+
+**Deprecated**
+
+
+2.6.0 (2021-05-10)
+------------------
+
+**Added**
+
+* Add json schema and wrapper for imaging data ``life.qbic.datamodel.datasets.imaging.ImageMetadata`` (`#199 <https://github.com/qbicsoftware/data-model-lib/pull/199>`_)
+
+* Provide new classes that describe incoming nf-core bioinformatic pipeline data structures and contains it in a dedicated NfCorePipelineResult class (`#198 <https://github.com/qbicsoftware/data-model-lib/pull/198>`_)
+
+**Fixed**
+
+**Dependencies**
+
+**Deprecated**
+
+
+2.5.0 (2021-04-26)
+------------------
 
 **Added**
 
@@ -18,12 +46,18 @@ This project adheres to `Semantic Versioning <https://semver.org/>`_.
 
 * New EnumFactory ``life.qbic.datamodel.dtos.business.ProductCategoryFactory`` (`#192 <https://github.com/qbicsoftware/data-model-lib/pull/192>`_)
 
+* New property experimentalDesign for ``life.qbic.datamodel.dtos.business.Offer``
+
+* Introduce new Enum ``life.qbic.datamodel.dtos.business.AffiliationLabel`` and new EnumFactory ``life.qbic.datamodel.dtos.business.AffiliationLabelFactory`` (`#194 <https://github.com/qbicsoftware/data-model-lib/pull/194>`_)
+
 **Fixed**
 
 * Override ``equals()`` method for ``life.qbic.datamodel.dtos.business.OfferId`` and
   ``life.qbic.datamodel.dtos.business.TomatoId`` properly
 
 * Remove @CompileStatic annotation from ProductId to enable builder pattern usage during compilation (`#186 <https://github.com/qbicsoftware/data-model-lib/issues/186>`_)
+
+* Override ``equals()`` method for ``life.qbic.datamodel.dtos.business.ProductId`` properly
 
 **Dependencies**